+++ scripts/config/menu.config.mjs
@@ -188,6 +188,8 @@ export class SimpleCoverVariantConfig extends HandlebarsApplicationMixin(Applica
                     this.#createSettingField(SETTING_KEYS.CREATURES_HALF_ONLY), 
                     this.#createSettingField(SETTING_KEYS.CREATURES_PRONE),
                     this.#createSettingField(SETTING_KEYS.IGNORE_DISTANCE_AOE),
+                    this.#createSettingField(SETTING_KEYS.IGNORE_ALL_AOE),
+                    this.#createSettingField(SETTING_KEYS.IGNORE_DISTANCE_SPACE),
                 ].filter(Boolean);
                 context.legend = game.i18n.localize("SIMPLE_COVER_5E.Settings.VariantMenu.Groups.General");
                 break;