+++ scripts/config/settings.config.mjs
@@ -62,6 +62,22 @@ const SETTINGS = [
     config: false
   },
   {
+    key: SETTING_KEYS.IGNORE_ALL_AOE,
+    name: "SIMPLE_COVER_5E.Settings.IgnoreAllAOE.Name",
+    hint: "SIMPLE_COVER_5E.Settings.IgnoreAllAOE.Hint",
+    type: new foundry.data.fields.BooleanField({ initial: false }),
+    requiresReload: false,
+    config: false
+  },
+  {
+    key: SETTING_KEYS.IGNORE_DISTANCE_SPACE,
+    name: "SIMPLE_COVER_5E.Settings.IgnoreDistanceSpace.Name",
+    hint: "SIMPLE_COVER_5E.Settings.IgnoreDistanceSpace.Hint",
+    type: new foundry.data.fields.BooleanField({ initial: false }),
+    requiresReload: false,
+    config: false
+  },
+  {
     key: SETTING_KEYS.CREATURES_PRONE,
     name: "SIMPLE_COVER_5E.Settings.CreaturesProne.Name",
     hint: "SIMPLE_COVER_5E.Settings.CreaturesProne.Hint",